requirements.txt
appdirs==1.4.4
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
Brotli==1.0.9
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
cryptography==36.0.1
cssselect2==0.4.1
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.6
django-cors-headers==3.8.0
django-filter==2.4.0
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-xml==2.0.0
eight==1.0.1
entrypoints==0.3
fonttools==4.28.4
future==0.18.2
html5lib==1.1
idna==3.3
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
lxml==4.7.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mysqlclient==2.0.3
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
pycparser==2.21
pydyf==0.1.2
Pygments==2.10.0
pymongo==3.12.0
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==3.0.6
pyphen==0.11.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.19.0
pytz==2021.1
pywin32==302
pywinpty==1.1.6
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.3
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
Send2Trash==1.8.0
signxml==2.9.0
six==1.16.0
sqlparse==0.4.1
terminado==0.12.1
testpath==0.5.0
tinycss2==1.1.1
tornado==6.1
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
weasyprint==53.4
webencodings==0.5.1
widgetsnbextension==3.5.2
xmlsec==1.3.9
zeep==4.0.0
zopfli==0.1.9
